Update iowin32.c
Proposed fix for issue "zlib 1.2.8 always uses Windows 8 - level API" #49.

The code is trying to define IOWIN32_USING_WINRT_API if and only if the target is a Metro app, by using the test:

#if WINAPI_FAMILY_PARTITION(WINAPI_PARTITION_APP)
However, that will be true for both desktop and Metro applications. To test for only Metro applications, and not desktop applications, the following will work:

#if !WINAPI_FAMILY_PARTITION(WINAPI_PARTITION_DESKTOP)